+++ examples/resources/routeros_capsman_access_list/import.sh
@@ -1,0 +1,3 @@
+#The ID can be found via API or the terminal
+#The command for the terminal is -> :put [/caps-man/access-list get [print show-ids]]
+terraform import routeros_capsman_access_list.test_rule "*1"