+++ routeros/mikrotik_serialize.go
@@ -283,7 +283,7 @@ func MikrotikResourceDataToTerraform(item MikrotikItem, s map[string]*schema.Sch
 	var diags diag.Diagnostics
 	var err error
 	var transformSet map[string]string
-	var setUnsetFields map[string]struct{}
+	var setUnsetFields, skipFields map[string]struct{}
 
 	// {"channel": "channel.config", "mikrotik-field-name": "schema-field-name"}
 	if ts, ok := s[MetaTransformSet]; ok {
@@ -293,6 +293,9 @@ func MikrotikResourceDataToTerraform(item MikrotikItem, s map[string]*schema.Sch
 	// "field_first", "field_second", "field_third"
 	if suf, ok := s[MetaSetUnsetFields]; ok {
 		setUnsetFields = loadSkipFields(suf.Default.(string))
+  }
+	if sf, ok := s[MetaSkipFields]; ok {
+		skipFields = loadSkipFields(sf.Default.(string))
 	}
 
 	// TypeMap,TypeSet initialization information storage.
@@ -324,6 +327,12 @@ func MikrotikResourceDataToTerraform(item MikrotikItem, s map[string]*schema.Sch
 
 		// field-name => field_name
 		terraformSnakeName := KebabToSnake(mikrotikKebabName)
+
+		if skipFields != nil {
+			if _, ok := skipFields[terraformSnakeName]; ok {
+				continue
+			}
+		}
 
 		// Composite fields.
 		var subFieldSnakeName string