+++ routeros/provider.go
@@ -103,6 +103,7 @@ func Provider() *schema.Provider {
 			"routeros_interface_bridge_port":     ResourceInterfaceBridgePort(),
 			"routeros_interface_bridge_vlan":     ResourceInterfaceBridgeVlan(),
 			"routeros_interface_bridge_settings": ResourceInterfaceBridgeSettings(),
+			"routeros_interface_eoip":            ResourceInterfaceEoip(),
 			"routeros_interface_gre":             ResourceInterfaceGre(),
 			"routeros_interface_vlan":            ResourceInterfaceVlan(),
 			"routeros_interface_vrrp":            ResourceInterfaceVrrp(),
@@ -144,6 +145,7 @@ func Provider() *schema.Provider {
 			// https://help.mikrotik.com/docs/display/ROS/WifiWave2#WifiWave2-WifiWave2CAPsMAN
 
 			// CAPsMAN Objects
+			"routeros_capsman_access_list":       ResourceCapsManAccessList(),
 			"routeros_capsman_channel":           ResourceCapsManChannel(),
 			"routeros_capsman_configuration":     ResourceCapsManConfiguration(),
 			"routeros_capsman_datapath":          ResourceCapsManDatapath(),